Fix issues with replying to RPC messages using Promises
Issue 1: If a message is received with the replyTo field set (an RPC
message) and the onMessage function that's passed to activateConsumer()
or startConsumer() returns a Promise and the Promise isn't an ES6
Promise, a JSON version of the Promise itself may be returned in the
reply message instead of the Promise's resolved value.
https://stackoverflow.com/questions/27746304/how-do-i-tell-if-an-object-is-a-promise

Issue 2: If a message is received with the replyTo field set (an RPC
message) and the onMessage function that's passed to startConsumer()
returns a Promise, the Promise's resolved value isn't returned in the
reply message. Instead a JSON version of the entire Promise is returned.