Make replying to a message with the replyTo field set optional
There are situations where a consumer may not want to immediately
reply to a message even though the message has the replyTo field set.
This change allows the consumer to decide whether or not to reply by
simply not returning a value or message if it doesn't want to reply.